At least partially use ssize_t for read(2)/write(2)

+++ ucspi-tcp/buffer_get.c
@@ -2,7 +2,7 @@
 #include "byte.h"
 #include "error.h"
 
-static int oneread(int (*op)(),int fd,char *buf,unsigned int len)
+static ssize_t oneread(ssize_t (*op)(),int fd,char *buf,unsigned int len)
 {
   int r;
 
@@ -27,7 +27,7 @@ int buffer_feed(buffer *s)
   int r;
 
   if (s->p) return s->p;
-  r = oneread(s->op,s->fd,s->x,s->n);
+  r = (int)oneread(s->op,s->fd,s->x,s->n);
   if (r <= 0) return r;
   s->p = r;
   s->n -= r;
@@ -40,7 +40,7 @@ int buffer_bget(buffer *s,char *buf,unsigned int len)
   int r;
  
   if (s->p > 0) return getthis(s,buf,len);
-  if (s->n <= len) return oneread(s->op,s->fd,buf,s->n);
+  if (s->n <= len) return (int)oneread(s->op,s->fd,buf,s->n);
   r = buffer_feed(s); if (r <= 0) return r;
   return getthis(s,buf,len);
 }
@@ -50,7 +50,7 @@ int buffer_get(buffer *s,char *buf,unsigned int len)
   int r;
  
   if (s->p > 0) return getthis(s,buf,len);
-  if (s->n <= len) return oneread(s->op,s->fd,buf,len);
+  if (s->n <= len) return (int)oneread(s->op,s->fd,buf,len);
   r = buffer_feed(s); if (r <= 0) return r;
   return getthis(s,buf,len);
 }
